Separate files for types that support or implement NodeRowObserver, NodeRowViewModel protocols
