+++ Stitch/Graph/Node/Port/ViewModel/NodeRowViewModel/NodeRowViewModel.swift
@@ -9,77 +9,6 @@ import Foundation
 import SwiftUI
 import StitchSchemaKit
 
-enum GraphItemType: Hashable {
-    case node(CanvasItemId)
-    
-    // Passing in layer input type ensures uniqueness of IDs in inspector
-    case layerInspector(NodeIOPortType) // portId (layer output) or layer-input-type (layer input)
-}
-
-extension GraphItemType {
-    static let empty: Self = .layerInspector(.keyPath(.init(layerInput: .size,
-                                                            portType: .packed)))
-    
-    var isLayerInspector: Bool {
-        switch self {
-        case .layerInspector:
-            return true
-        default:
-            return false
-        }
-    }
-    
-    var layerInputPort: LayerInputPort? {
-        switch self {
-        case .node(let canvasItemId):
-            return canvasItemId.layerInputCase?.keyPath.layerInput
-        case .layerInspector(let nodeIOPortType):
-            return nodeIOPortType.keyPath?.layerInput
-        }
-    }
-    
-    var getLayerInputCoordinateOnGraph: LayerInputCoordinate? {
-        switch self {
-        case .node(let x):
-            switch x {
-            case .layerInput(let x):
-                return x
-            default:
-                return nil
-            }
-        default:
-            return nil
-        }
-    }
-}
-
-struct NodeRowViewModelId: Hashable {
-    var graphItemType: GraphItemType
-    var nodeId: NodeId
-    
-    // TODO: this is always 0 for layer inspector which creates issues for tabbing
-    var portId: Int
-}
-
-extension NodeRowViewModelId {
-    /// Determines if some row view model reports to a node, rather than to the layer inspector
-    var isNode: Bool {
-        switch self.graphItemType {
-        case .node:
-            return true
-        default:
-            return false
-        }
-    }
-    
-    static let empty: Self = .init(graphItemType: .node(.node(.init())),
-                                   nodeId: .init(),
-                                   portId: -1)
-    
-    var layerInputPort: LayerInputPort? {
-        self.graphItemType.layerInputPort
-    }
-}
 
 protocol NodeRowViewModel: StitchLayoutCachable, Observable, Identifiable {
     associatedtype FieldType: FieldViewModel where FieldType.NodeRowType == Self
@@ -280,154 +209,6 @@ extension PortValue {
     }
 }
 
-// UI data
-@Observable
-final class InputNodeRowViewModel: NodeRowViewModel {
-    typealias PortViewType = InputPortViewData
-    
-    static let nodeIO: NodeIO = .input
-    
-    let id: NodeRowViewModelId
-    @MainActor var viewCache: NodeLayoutCache?
-    @MainActor var activeValue: PortValue = .number(.zero)
-    @MainActor var fieldValueTypes = FieldGroupTypeDataList<InputFieldViewModel>()
-    @MainActor var connectedCanvasItems: Set<CanvasItemId> = .init()
-    @MainActor var anchorPoint: CGPoint?
-    @MainActor var portColor: PortColor = .noEdge
-    @MainActor var isDragging = false
-    @MainActor var portViewData: PortViewType?
-    @MainActor weak var nodeDelegate: NodeDelegate?
-    @MainActor weak var rowDelegate: InputNodeRowObserver?
-    
-    // TODO: input node row view model for an inspector should NEVER have canvasItemDelegate
-    @MainActor weak var canvasItemDelegate: CanvasItemViewModel? // also nil when the layer input is not on the canvas
-    
-    // TODO: temporary property for old-style layer nodes
-    @MainActor var layerPortId: Int?
-    
-    @MainActor
-    init(id: NodeRowViewModelId,
-         rowDelegate: InputNodeRowObserver?,
-         canvasItemDelegate: CanvasItemViewModel?) {
-        self.id = id
-        self.nodeDelegate = nodeDelegate
-        self.rowDelegate = rowDelegate
-        self.canvasItemDelegate = canvasItemDelegate
-    }
-}
-
-extension InputNodeRowViewModel {
-    @MainActor
-    func findConnectedCanvasItems() -> CanvasItemIdSet {
-        guard let upstreamOutputObserver = self.rowDelegate?.upstreamOutputObserver,
-              let upstreamNodeRowViewModel = upstreamOutputObserver.nodeRowViewModel,
-              let upstreamId = upstreamNodeRowViewModel.canvasItemDelegate?.id else {
-            return .init()
-        }
-        
-        return Set([upstreamId])
-    }
-    
-    @MainActor
-    func calculatePortColor() -> PortColor {
-        let isEdgeSelected = self.hasSelectedEdge()
-        
-        // Note: inputs always ignore actively-drawn or animated (edge-edit-mode) edges etc.
-        let isSelected = self.isCanvasItemSelected ||
-            self.isConnectedToASelectedCanvasItem ||
-            isEdgeSelected
-        return .init(isSelected: isSelected,
-                     hasEdge: hasEdge,
-                     hasLoop: hasLoop)
-    }
-    
-    @MainActor 
-    func hasSelectedEdge() -> Bool {
-        guard let portViewData = portViewData,
-              let graphDelegate = graphDelegate else {
-            return false
-        }
-        
-        return graphDelegate.selectedEdges.contains { $0.to == portViewData }
-    }
-}
-
-@Observable
-final class OutputNodeRowViewModel: NodeRowViewModel {
-    typealias PortViewType = OutputPortViewData
-    static let nodeIO: NodeIO = .output
-    
-    let id: NodeRowViewModelId
-    @MainActor var viewCache: NodeLayoutCache?
-    @MainActor var activeValue: PortValue = .number(.zero)
-    @MainActor var fieldValueTypes = FieldGroupTypeDataList<OutputFieldViewModel>()
-    @MainActor var connectedCanvasItems: Set<CanvasItemId> = .init()
-    @MainActor var anchorPoint: CGPoint?
-    @MainActor var portColor: PortColor = .noEdge
-    @MainActor var isDragging = false
-    @MainActor var portViewData: PortViewType?
-    @MainActor weak var nodeDelegate: NodeDelegate?
-    @MainActor weak var rowDelegate: OutputNodeRowObserver?
-    @MainActor weak var canvasItemDelegate: CanvasItemViewModel?
-    
-    init(id: NodeRowViewModelId,
-         rowDelegate: OutputNodeRowObserver?,
-         canvasItemDelegate: CanvasItemViewModel?) {
-        self.id = id
-        self.nodeDelegate = nodeDelegate
-        self.rowDelegate = rowDelegate
-        self.canvasItemDelegate = canvasItemDelegate
-    }
-}
-
-extension OutputNodeRowViewModel {
-    @MainActor
-    func findConnectedCanvasItems() -> CanvasItemIdSet {
-        guard let downstreamCanvases = self.rowDelegate?.getConnectedDownstreamNodes() else {
-            return .init()
-        }
-            
-        let downstreamCanvasIds = downstreamCanvases.map { $0.id }
-        return Set(downstreamCanvasIds)
-    }
-    
-    /// Note: an actively-drawn edge SITS ON TOP OF existing edges. So there is no distinction between port color vs edge color.
-    /// An actively-drawn edge's color is determined only by:
-    /// 1. "Do we have a loop?" (blue vs theme-color) and
-    /// 2. "Do we have an eligible input?" (highlight vs non-highlighted)
-    @MainActor
-    func calculatePortColor() -> PortColor {
-        if let drawingObserver = self.graphDelegate?.edgeDrawingObserver,
-           let drawnEdge = drawingObserver.drawingGesture,
-           drawnEdge.output.id == self.id {
-            return PortColor(
-                isSelected: drawingObserver.nearestEligibleInput.isDefined,
-//                hasEdge: hasEdge,
-                hasEdge: drawingObserver.nearestEligibleInput.isDefined,
-                hasLoop: hasLoop)
-        }
-        
-        
-        // Otherwise, common port color logic applies:
-        else {
-            let isSelected = self.isCanvasItemSelected ||
-                self.isConnectedToASelectedCanvasItem ||
-            self.hasSelectedEdge()
-            return PortColor(isSelected: isSelected,
-                             hasEdge: hasEdge,
-                             hasLoop: hasLoop)
-        }
-    }
-    
-    @MainActor func hasSelectedEdge() -> Bool {
-        guard let portViewData = portViewData,
-              let graphDelegate = graphDelegate else {
-            return false
-        }
-        
-        return graphDelegate.selectedEdges.contains { $0.from == portViewData }
-    }
-}
 
 extension CanvasItemViewModel {
     // TODO: This is impossible to find via a code search; too many methods are called `sync`
